+++ examples/bbm_1d/bbm_1d_fourier.jl
@@ -5,7 +5,7 @@ using SummationByPartsOperators: fourier_derivative_operator
 ###############################################################################
 # Semidiscretization of the BBM equation
 
-equations = BBMEquation1D(gravity_constant = 9.81)
+equations = BBMEquation1D(gravity = 9.81)
 
 initial_condition = initial_condition_convergence_test
 boundary_conditions = boundary_condition_periodic