+++ examples/bbm_bbm_1d/bbm_bbm_1d_dg.jl
@@ -9,7 +9,7 @@ using DispersiveShallowWater
 # Semidiscretization of the BBM-BBM equations
 
 # or bathymetry_variable instead of bathymetry_flat
-equations = BBMBBMEquations1D(bathymetry_type = bathymetry_flat, gravity_constant = 1.0)
+equations = BBMBBMEquations1D(bathymetry_type = bathymetry_flat, gravity = 1.0)
 
 # initial_condition_convergence_test needs periodic boundary conditions
 initial_condition = initial_condition_convergence_test