+++ examples/serre_green_naghdi_1d/serre_green_naghdi_soliton.jl
@@ -5,7 +5,7 @@ using DispersiveShallowWater
 # Semidiscretization of the Serre-Green-Naghdi equations
 
 equations = SerreGreenNaghdiEquations1D(bathymetry_type = bathymetry_flat,
-                                        gravity_constant = 9.81)
+                                        gravity = 9.81)
 
 # initial_condition_convergence_test needs periodic boundary conditions
 initial_condition = initial_condition_convergence_test