+++ src/equations/bbm_1d.jl
@@ -1,5 +1,5 @@
 @doc raw"""
-    BBMEquation1D(; gravity_constant, D = 1.0, eta0 = 0.0, split_form = true)
+    BBMEquation1D(; gravity, D = 1.0, eta0 = 0.0, split_form = true)
 
 BBM (Benjamin–Bona–Mahony) equation in one spatial dimension.
 The equation is given by
@@ -10,7 +10,7 @@ The equation is given by
 ```
 
 The unknown quantity of the BBM equation is the total water height ``\eta``.
-The gravitational constant is denoted by `g` and the constant bottom topography (bathymetry) ``b = \eta_0 - D``,
+The gravitational acceleration `gravity` is denoted by ``g`` and the constant bottom topography (bathymetry) ``b = \eta_0 - D``,
 where ``\eta_0`` is the constant still-water surface and ``D`` the still-water depth. The water height above
 the bathymetry is therefore given by ``h = \eta - \eta_0 + D``. The BBM equation is only implemented for ``\eta_0 = 0``.
 
@@ -42,15 +42,15 @@ for periodic boundary conditions.
   [DOI: 10.1007/s10543-023-00992-w](https://doi.org/10.1007/s10543-023-00992-w)
 """
 struct BBMEquation1D{RealT <: Real} <: AbstractBBMEquation{1, 1}
-    gravity::RealT # gravitational constant
+    gravity::RealT # gravitational acceleration
     D::RealT # still-water depth
     eta0::RealT # constant still-water surface
     split_form::Bool # whether to use a split-form or not
 end
 
-function BBMEquation1D(; gravity_constant, D = 1.0, eta0 = 0.0, split_form = true)
+function BBMEquation1D(; gravity, D = 1.0, eta0 = 0.0, split_form = true)
     eta0 == 0.0 || @warn "The still-water surface needs to be 0 for the BBM equations"
-    BBMEquation1D(gravity_constant, D, eta0, split_form)
+    BBMEquation1D(gravity, D, eta0, split_form)
 end
 
 """
@@ -65,7 +65,7 @@ See section 4.1.3 in (there is an error in paper: it should be `sech^2` instead 
   [DOI: 10.4208/cicp.OA-2020-0119](https://doi.org/10.4208/cicp.OA-2020-0119)
 """
 function initial_condition_convergence_test(x, t, equations::BBMEquation1D, mesh)
-    g = gravity_constant(equations)
+    g = gravity(equations)
     D = equations.D
     alpha = sqrt(g * D)
     beta = 3 / 2 * sqrt(g / D)
@@ -96,7 +96,7 @@ end
 A smooth manufactured solution in combination with [`initial_condition_manufactured`](@ref).
 """
 function source_terms_manufactured(q, x, t, equations::BBMEquation1D)
-    g = gravity_constant(equations)
+    g = gravity(equations)
     D = still_waterdepth(q, equations)
     a1 = sqrt(g * D)
     a2 = sqrt(g / D)
@@ -121,7 +121,7 @@ function create_cache(mesh, equations::BBMEquation1D,
     eta_x = zero(eta2)
     etaeta_x = zero(eta2)
     eta_xx = zero(eta2)
-    g = gravity_constant(equations)
+    g = gravity(equations)
     c_0 = sqrt(g * D)
     c_1 = sqrt(g / D)
     cache = (; invImD2, eta2, eta2_x, eta_x, etaeta_x, eta_xx, c_0, c_1,